Rename appdistribution directory to apptesting
